board 페이지 라우터 삭제
