+++ frontend/src/components/TokenCheck.jsx
@@ -10,7 +10,7 @@ export const TokenCheckProvider = ({ children }) => {
 
   const [isLogin, setIsLogin] = useState(false);
   const [userId, setUserId] = useState(null);
-
+  const [auth, setAuth] = useState(null);
 
   useEffect(() => {
     const accessToken = localStorage.getItem('accessToken');
@@ -42,6 +42,7 @@ export const TokenCheckProvider = ({ children }) => {
         .then((data) => {
           if (data) {
             setUserId(data.id);
+            setAuth(data.authority);
           }
         })
         .catch((error) => {
@@ -54,6 +55,7 @@ export const TokenCheckProvider = ({ children }) => {
   const UserTokenValue = {
     isLogin,
     userId,
+    auth,
   };
